requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
Jinja2==3.1.2
junos-eznc==2.6.7
lxml==4.9.2
MarkupSafe==2.1.2
ncclient==0.6.13
netaddr==0.8.0
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PyYAML==6.0
scp==0.14.5
six==1.16.0
transitions==0.9.0
yamlordereddictloader==0.4.0
